1. 修改gitconfig配置 2. 添加ssh config 配置的backup

+++ scripts/backup.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 cd "$(dirname $0)"/..
-sudo_cp_if_exists(){
-    if [ -f "$1" ];
+sudo_cp_if_file_exists(){
+    if [ -f "$1" -a -e $2 ];
     then
         sudo cp "$1" "$2"
     fi
@@ -14,9 +14,25 @@ cp_if_file_exists(){
         cp "$1" "$2"
     fi
 }
-cp ~/.config/nvim/init.vim neovim
-cp ~/.zshrc zsh
-cp ~/.tmux.conf tmux
-source ss/proxy.zsh
-proxy stop && cp_if_file_exists ~/.gitconfig git && cp_if_file_exists ~/.gitmessage git && proxy start
 
+local_backup(){
+
+	# vim 
+	cp ~/.config/nvim/init.vim neovim
+
+	# zsh
+	cp ~/.zshrc zsh
+
+	# tmux
+	cp ~/.tmux.conf tmux
+
+	# git 备份之前关闭了代理
+	source ss/proxy.zsh
+	proxy stop && cp_if_file_exists ~/.gitconfig git && cp_if_file_exists ~/.gitmessage git && proxy start
+
+	# ssh
+	cp ~/.ssh/config ssh
+}
+
+
+local_backup
